Revert "dev: use dep replace godep"
This reverts commit efc95a984da60c91e8f4a3ce623b08184727bd92.

+++ pkg/proxy/cfg.go
@@ -18,7 +18,7 @@ type Cfg struct {
 	PDAddrs       []string `json:"pdAddrs"`
 	MaxRetries    int      `json:"maxRetries"`
 	RetryDuration int64    `json:"retryDuration"`
-	WorkerCount   uint64   `json:"workerCount"`
+	WorkerCount   int      `json:"workerCount"`
 	SupportCMDs   []string `json:"supportCMDs"`
 }
